requirements.txt
aiosmtplib==2.0.2
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
Authlib==1.3.0
autoflake==2.3.1
autopep8==2.3.1
bcrypt==4.1.2
blinker==1.7.0
boto3==1.34.143
botocore==1.34.143
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colorlog==6.8.2
comm==0.2.2
cryptography==42.0.5
cssselect==1.2.0
cssutils==2.10.2
debugpy==1.8.1
decorator==5.1.1
distro==1.9.0
dnspython==2.6.1
ecdsa==0.18.0
email_validator==2.1.1
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.110.0
fastapi-mail==1.4.1
filetype==1.2.0
google-auth==2.29.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==3.6
importlib_metadata==7.1.0
ipykernel==6.29.4
ipython==8.18.1
ipython-genutils==0.2.0
ipython-sql==0.5.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
jupyter_client==8.6.1
jupyter_core==5.7.2
loguru==0.7.2
lxml==5.2.1
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==2.0.1
openai==1.20.0
orjson==3.9.15
packaging==24.0
pandas==2.2.2
parso==0.8.3
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.2.0
premailer==3.10.0
prettytable==3.10.0
prompt-toolkit==3.0.43
psutil==5.9.8
psycopg2-binary==2.9.9
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1_modules==0.4.0
pycodestyle==2.12.0
pycparser==2.21
pydantic==2.6.4
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
pydantic_core==2.16.3
pyflakes==3.2.0
Pygments==2.17.2
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-http-client==3.3.7
python-jose==3.3.0
python-multipart==0.0.9
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
requests==2.31.0
rsa==4.9
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.28
sqlparse==0.4.4
stack-data==0.6.3
starkbank-ecdsa==2.2.0
starlette==0.36.3
text-unidecode==1.3
tokenize-rt==5.2.0
tomli==2.0.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
typing_extensions==4.10.0
tzdata==2024.1
ujson==5.9.0
urllib3==1.26.19
uvicorn==0.28.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
win32-setctime==1.1.0
zipp==3.18.1
